Added Enum.hash
Before this commit, the hash method was not overriden. Therefore, ruby
gives the objects random hashcodes which are different from the value's hash.
Also, a call to new will assign a different hashcode to the object even if
new is overriden and does not create any new instances.

This commit ensures that the hash code of enum members is the same as the
hash code as their underlying values.